Deleted test containers after test completion (#521)
* Deleted test containers after test completion

* Merged RoleRepositoryCollectionIT in IntegerIdDomainRepositoryIT

* Fixed codacy issues
