+++ src/test/java/com/microsoft/azure/spring/data/cosmosdb/repository/integration/ContactRepositoryIT.java
@@ -13,6 +13,7 @@ import com.microsoft.azure.spring.data.cosmosdb.repository.repository.ContactRep
 import com.microsoft.azure.spring.data.cosmosdb.repository.support.CosmosEntityInformation;
 import org.assertj.core.util.Lists;
 import org.junit.After;
+import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Ignore;
@@ -35,8 +36,11 @@ public class ContactRepositoryIT {
 
     private static final Contact TEST_CONTACT = new Contact("testId", "faketitle");
 
-    private final CosmosEntityInformation<Contact, String> entityInformation =
+    private static final CosmosEntityInformation<Contact, String> entityInformation =
             new CosmosEntityInformation<>(Contact.class);
+
+    private static CosmosTemplate staticTemplate;
+    private static boolean isSetupDone;
 
     @Autowired
     ContactRepository repository;
@@ -46,12 +50,22 @@ public class ContactRepositoryIT {
 
     @Before
     public void setup() {
+        if (!isSetupDone) {
+            staticTemplate = template;
+            template.createContainerIfNotExists(entityInformation);
+        }
         repository.save(TEST_CONTACT);
+        isSetupDone = true;
     }
 
     @After
     public void cleanup() {
         repository.deleteAll();
+    }
+
+    @AfterClass
+    public static void afterClassCleanup() {
+        staticTemplate.deleteContainer(entityInformation.getContainerName());
     }
 
     @Test
